Argument fixes.
git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@6786 e44c5d27-f830-0410-8ce8-974f7a764be6

+++ src/user/cs_user_metal_structures_source_terms.f90
@@ -49,6 +49,7 @@ use cs_c_bindings
 use cs_f_interfaces
 
 use cs_tagms
+
 !===============================================================================
 
 implicit none
@@ -62,7 +63,7 @@ integer          izmet, met_znb
 integer          ltmast(ncelet)
 integer          izmast(ncelet)
 
-double precision itypst(ncelet,nvar)
+integer          itypst(ncelet,nvar)
 double precision svcond(ncelet,nvar)
 double precision tmet
